+++ apps/restconf/clixon_restconf.h
@@ -53,8 +53,8 @@ enum restconf_media{
     YANG_DATA_XML,       /* "application/yang-data+xml" */
     YANG_PATCH_JSON,     /* "application/yang-patch+json" */
     YANG_PATCH_XML,      /* "application/yang-patch+xml" */
-    YANG_COLLECTION_XML, /* draft-ietf-netconf-restconf-collection-00.txt */
-    YANG_COLLECTION_JSON /* draft-ietf-netconf-restconf-collection-00.txt */
+    YANG_COLLECTION_XML, /* draft-wwlh-netconf-list-pagination-rc-01.txt */
+    YANG_COLLECTION_JSON /* draft-wwlh-netconf-list-pagination-rc-01.txt */
 };
 typedef enum restconf_media restconf_media;