+++ include/clixon_custom.h
@@ -124,7 +124,3 @@
  * draft-wwlh-netconf-list-pagination-rc-01
  */
 #define LIST_PAGINATION
-
-/*! Clixon netconf pagination
- */
-#define CLIXON_PAGINATION