+++ lib/clixon/clixon_xml.h
@@ -55,13 +55,17 @@
 #define NETCONF_INPUT_CONFIG "config"
 
 /* Collections namespace from draft-ietf-netconf-restconf-collection-00.txt
+ * XXX: Obsolete but may come back in style
  */
 #define NETCONF_COLLECTION_NAMESPACE "urn:ietf:params:xml:ns:yang:ietf-netconf-list-pagination"
 
-/* See RFC 7950 Sec 5.3.1: YANG defines an XML namespace for NETCONF <edit-config> 
- * operations, <error-info> content, and the <action> element.
+/* Collections namespace for Clixon
  */
-#define NETCONF_COLLECTION_NAMESPACE "urn:ietf:params:xml:ns:yang:ietf-netconf-list-pagination"
+#define CLIXON_PAGINATON_NAMESPACE "http://clicon.org/clixon-netconf-list-pagination"
+
+/* Collections namespace for restconf
+ */
+#define RESTCONF_PAGINATON_NAMESPACE "urn:ietf:params:xml:ns:yang:ietf-restconf-list-pagination"
 
 /* Output symbol for netconf get/get-config
  *    ietf-netconf.yang defines it as output: